index u128

+++ runtime/src/installed_scheduler_pool.rs
@@ -48,7 +48,7 @@ pub fn initialized_result_with_timings() -> ResultWithTimings {
     (Ok(()), ExecuteTimings::default())
 }
 
-pub type Index = u64;
+pub type Index = u128;
 
 pub trait InstalledSchedulerPool: Send + Sync + Debug {
     fn take_scheduler(&self, context: SchedulingContext) -> InstalledSchedulerBox {
